+++ CRM/CiviGeometry/Upgrader.php
@@ -3,7 +3,7 @@
 /**
  * Collection of upgrade steps.
  */
-class CRM_CiviGeometry_Upgrader extends CRM_CiviGeometry_Upgrader_Base {
+class CRM_CiviGeometry_Upgrader extends CRM_Extension_Upgrader_Base {
 
   // By convention, functions that look like "function upgrade_NNNN()" are
   // upgrade tasks. They are executed in order (like Drupal's hook_update_N).
@@ -165,6 +165,24 @@ end
     CRM_Core_DAO::executeQuery("ALTER TABLE civigeometry_geometry_collection_type ADD UNIQUE INDEX index_label(label)");
     CRM_Core_DAO::executeQuery("ALTER TABLE civigeometry_geometry DROP INDEX index_geometry_type_label_is_archived");
     CRM_Core_DAO::executeQuery("ALTER TABLE civigeometry_geometry ADD INDEX index_is_archived_geometry_type_label(is_archived,geometry_type_id,label)");
+    return TRUE;
+  }
+
+  /**
+   * Add index to geometry_entity table
+   * (using update number 5184 to reflect Civi major version and extension major/minor/patch)
+   * @return TRUE on success
+   * @throws Exception
+   */
+  public function upgrade_5184() {
+    $this->ctx->log->info('Applying update 5184 - Adding index on civigeometry_geometry_entity table');
+    CRM_Core_DAO::executeQuery("ALTER TABLE civigeometry_geometry_entity ADD INDEX index_entity_table_entity_id(entity_table,entity_id)");
+    return TRUE;
+  }
+
+  public function upgrade_5200() {
+    $this->ctx->log->info('Applying update 5200 - Adding reason field to civigeometry_geometry_entity table');
+    CRM_Core_DAO::executeQuery("ALTER TABLE civigeometry_geometry_entity ADD COLUMN reason varchar(64) DEFAULT NULL COMMENT 'Reason for relationship between geometry and entity'");
     return TRUE;
   }