+++ Civi/Api4/GeometryEntity.php
@@ -2,20 +2,19 @@
 namespace Civi\Api4;
 
 /**
- * GeometryType entity.
+ * GeometryEntity entity.
  *
  * Provided by the CiviGeometry extension.
  *
  * @package Civi\Api4
  */
-class GeometryType extends Generic\DAOEntity {
+class GeometryEntity extends Generic\DAOEntity {
 
   public static function permissions() {
     return [
+      'default' => ['access geometry'],
       'create' => [['administer geometry', 'administer civicrm']],
       'delete' => [['administer geometry', 'administer civicrm']],
-      'default' => ['access geometry'],
     ];
   }
-
 }