+++ civigeometry.php
@@ -14,15 +14,6 @@ function civigeometry_civicrm_config(&$config) {
 }
 
 /**
- * Implements hook_civicrm_xmlMenu().
- *
- * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_xmlMenu
- */
-function civigeometry_civicrm_xmlMenu(&$files) {
-  _civigeometry_civix_civicrm_xmlMenu($files);
-}
-
-/**
  * Implements hook_civicrm_install().
  *
  * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_install
@@ -32,107 +23,12 @@ function civigeometry_civicrm_install() {
 }
 
 /**
- * Implements hook_civicrm_postInstall().
- *
- * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_postInstall
- */
-function civigeometry_civicrm_postInstall() {
-  _civigeometry_civix_civicrm_postInstall();
-}
-
-/**
- * Implements hook_civicrm_uninstall().
- *
- * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_uninstall
- */
-function civigeometry_civicrm_uninstall() {
-  _civigeometry_civix_civicrm_uninstall();
-}
-
-/**
  * Implements hook_civicrm_enable().
  *
  * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_enable
  */
 function civigeometry_civicrm_enable() {
   _civigeometry_civix_civicrm_enable();
-}
-
-/**
- * Implements hook_civicrm_disable().
- *
- * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_disable
- */
-function civigeometry_civicrm_disable() {
-  _civigeometry_civix_civicrm_disable();
-}
-
-/**
- * Implements hook_civicrm_upgrade().
- *
- * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_upgrade
- */
-function civigeometry_civicrm_upgrade($op, CRM_Queue_Queue $queue = NULL) {
-  return _civigeometry_civix_civicrm_upgrade($op, $queue);
-}
-
-/**
- * Implements hook_civicrm_managed().
- *
- * Generate a list of entities to create/deactivate/delete when this module
- * is installed, disabled, uninstalled.
- *
- * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_managed
- */
-function civigeometry_civicrm_managed(&$entities) {
-  _civigeometry_civix_civicrm_managed($entities);
-}
-
-/**
- * Implements hook_civicrm_caseTypes().
- *
- * Generate a list of case-types.
- *
- * Note: This hook only runs in CiviCRM 4.4+.
- *
- * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_caseTypes
- */
-function civigeometry_civicrm_caseTypes(&$caseTypes) {
-  _civigeometry_civix_civicrm_caseTypes($caseTypes);
-}
-
-/**
- * Implements hook_civicrm_angularModules().
- *
- * Generate a list of Angular modules.
- *
- * Note: This hook only runs in CiviCRM 4.5+. It may
- * use features only available in v4.6+.
- *
- * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_angularModules
- */
-function civigeometry_civicrm_angularModules(&$angularModules) {
-  _civigeometry_civix_civicrm_angularModules($angularModules);
-}
-
-/**
- * Implements hook_civicrm_alterSettingsFolders().
- *
- * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_alterSettingsFolders
- */
-function civigeometry_civicrm_alterSettingsFolders(&$metaDataFolders = NULL) {
-  _civigeometry_civix_civicrm_alterSettingsFolders($metaDataFolders);
-}
-
-/**
- * Implements hook_civicrm_entityTypes().
- *
- * Declare entity types provided by this module.
- *
- * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_entityTypes
- */
-function civigeometry_civicrm_entityTypes(&$entityTypes) {
-  _civigeometry_civix_civicrm_entityTypes($entityTypes);
 }
 
 /**
@@ -162,14 +58,64 @@ function civigeometry_civicrm_permission(&$permissions) {
  * @link https://docs.civicrm.org/dev/en/latest/hooks/hook_civicrm_permission/
  */
 function civigeometry_civicrm_alterAPIPermissions($entity, $action, &$params, &$permissions) {
-  $permissions['geometry']['create'] = $permissions['geometry']['delete'] = array(array('administer geometry', 'administer civicrm'));
+  $permissions['address']['getgeometries'] = [['access civicrm', 'access AJAX API']];
+  $permissions['geometry']['create'] = $permissions['geometry']['delete'] = [['administer geometry', 'administer civicrm']];
   $permissions['geometry']['default'] = array('access geometry');
-  $permissions['geometry_collection']['create'] = array(array('administer geometry', 'administer civicrm'));
-  $permissions['geometry_collection']['default'] = array('access geometry');
+  $permissions['geometry_collection']['create'] = [['administer geometry', 'administer civicrm']];
+  $permissions['geometry_collection']['default'] = ['access geometry'];
   $permissions['geometry_collection']['unarchive'] = $permissions['geometry_collection']['archive'] = $permissions['geometry_collection']['delete'] = $permissions['geometry_collection']['create'];
-  $permissions['geometry_type']['create'] = $permissions['geometry_type']['delete'] = array(array('administer geometry', 'administer civicrm'));
-  $permissions['geometry_type']['default'] = array('access geometry');
+  $permissions['geometry_type']['create'] = $permissions['geometry_type']['delete'] = [['administer geometry', 'administer civicrm']];
+  $permissions['geometry_type']['default'] = ['access geometry'];
   $permissions['geometry_collection_type'] = $permissions['geometry_type'];
+}
+
+/**
+ *
+ * Callback function for enqueuing new geoplaceAddress tasks
+ *
+ * @param int $objectId
+ *
+ */
+function _civigeometry_geoplaceAddress($objectId) {
+  $queue = CRM_CiviGeometry_Helper::singleton()->getQueue();
+  $address = civicrm_api3('Address', 'get', ['id' => $objectId])['values'][$objectId];
+  if (!empty($address['geo_code_2']) && !empty($address['geo_code_1'])) {
+    $task = new CRM_Queue_Task(
+      ['CRM_CiviGeometry_Tasks', 'geoplaceAddress'],
+      [$objectId]
+    );
+    $queue->createItem($task);
+  }
+}
+
+/**
+ *
+ * Callback function for enqueuing new archive geometry cleanup tasks
+ *
+ * @param int $objectId
+ *
+ */
+function _civigeometry_archiveGeometry($objectId) {
+  $dao = new CRM_CiviGeometry_DAO_GeometryEntity();
+  $dao->whereAdd(CRM_Core_DAO::composeQuery('geometry_id = %1', [1 => [$objectId, 'Positive']]));
+  $dao->whereAdd("entity_table = 'civicrm_address'");
+  $dao->delete(TRUE);
+}
+
+/**
+ *
+ * Callback function for enqueuing new geometry relationships tasks
+ *
+ * @param int $objectId
+ *
+ */
+function _civigeometry_buildGeometryRelationships($objectId) {
+  $queue = CRM_CiviGeometry_Helper::singleton()->getQueue();
+  $task = new CRM_Queue_Task(
+      ['CRM_CiviGeometry_Tasks', 'buildGeometryRelationships'],
+      [$objectId]
+  );
+  $queue->createItem($task);
 }
 
 /**
@@ -190,36 +136,35 @@ function civigeometry_symfony_civicrm_post($event) {
   // 2 = objectId
   // 3 = objectREf
   if ($hookValues[0] !== 'delete' && $hookValues[0] !== 'geoplace' && $hookValues[1] == 'Address') {
-    $queue = CRM_CiviGeometry_Helper::singleton()->getQueue();
-    $id = $hookValues[2];
-    $address = civicrm_api3('Address', 'get', ['id' => $id])['values'][$id];
-    if (!empty($address['geo_code_2']) && !empty($address['geo_code_1'])) {
-      $task = new CRM_Queue_Task(
-        ['CRM_CiviGeometry_Tasks', 'geoplaceAddress'],
-        [$id]
+    if (CRM_Core_Transaction::isActive()) {
+      CRM_Core_Transaction::addCallback(CRM_Core_Transaction::PHASE_POST_COMMIT,
+        '_civigeometry_geoplaceAddress', [$hookValues[2]]
       );
-      $queue->createItem($task);
+    }
+    else {
+      _civigeometry_geoplaceAddress($hookValues[2]);
     }
   }
   // If a geometry has been archived ensure that all address records of it in the GeometryEntity table are removed.
   if ($hookValues[0] == 'archive' && $hookValues[1] == 'Geometry') {
-    $id = $hookValues[2];
-    $dao = new CRM_CiviGeometry_DAO_GeometryEntity();
-    $dao->geometry_id = $id;
-    $dao->entity_table = 'civicrm_address';
-    if ($dao->find()) {
-      while ($dao->fetch()) {
-        $dao->delete();
-      }
+    if (CRM_Core_Transaction::isActive()) {
+      CRM_Core_Transaction::addCallback(CRM_Core_Transaction::PHASE_POST_COMMIT,
+        '_civigeometry_archiveGeometry', [$hookValues[2]]
+      );
+    }
+    else {
+      _civigeometry_archiveGeometry($hookValues[2]);
     }
   }
   if ($hookValues[0] == 'create' && $hookValues[1] == 'Geometry') {
-    $queue = CRM_CiviGeometry_Helper::singleton()->getQueue();
-    $task = new CRM_Queue_Task(
-        ['CRM_CiviGeometry_Tasks', 'buildGeometryRelationships'],
-        [$hookValues[2]]
-    );
-    $queue->createItem($task);
+    if (CRM_Core_Transaction::isActive()) {
+      CRM_Core_Transaction::addCallback(CRM_Core_Transaction::PHASE_POST_COMMIT,
+        '_civigeometry_buildGeometryRelationships', [$hookValues[2]]
+      );
+    }
+    else {
+      _civigeometry_buildGeometryRelationships($hookValues[2]);
+    }
   }
 }
 
@@ -235,7 +180,9 @@ function civigeometry_civicrm_merge($type, &$data, $mainId = NULL, $otherId = NU
       break;
 
     case 'sqls':
-      $data[] = "DELETE FROM civigeometry_geometry_entity WHERE entity_id NOT IN (SELECT id FROM civicrm_address) AND entity_table = 'civicrm_address'";
+      $data[] = "DELETE cge FROM civigeometry_geometry_entity cge
+                 INNER JOIN civicrm_address ca ON ca.id = cge.entity_id AND cge.entity_table = 'civicrm_address'
+                 WHERE ca.contact_id = {$otherId}";
       break;
 
   }